website/requirements.txt
django==1.9
django-reversion==2.0.0
django-crispy-forms==1.6.1
django-formtools==1.0
httplib2==0.9.2
xadmin==0.6.1
django-debug-toolbar==1.6
djangorestframework==3.5.1
django-filter==1.0.1
django-extensions==1.7.5
pytz==2016.10
django-qiniu-storage==2.3.0
social-auth-app-django==1.0.0
django-allauth==0.30.0
Pillow==4.0.0
django-grappelli==2.8.3
MySQL-python==1.2.5
django-redis=4.8.0
django-redis-sessions=0.5.6